move passport code and create login controller
