+++ server/app.ts
@@ -10,14 +10,13 @@ import { connect } from 'mongoose';
 import path from 'path';
 import SparkPost from 'sparkpost';
 import passport from 'passport';
-import { Strategy as FacebookStrategy } from 'passport-facebook';
-import { OAuth2Strategy as GoogleStrategy } from 'passport-google-oauth';
 
+import './passport';
 import config from './config';
-import { Maker, Request, User, UserLogin } from './schemas';
+import { Maker, Request } from './schemas';
 import { IMaker } from './interfaces';
-import { RequestsController } from './controllers';
-import { ensureAuthenticated, getIp, getUser } from './utils';
+import { LoginController, RequestsController } from './controllers';
+import { ensureAuthenticated, getUser } from './utils';
 
 const portNumber = process.env.PORT || 3050;
 const cookieSession: SessionOptions = {
@@ -49,136 +48,6 @@ connect(
   }
 );
 
-passport.use(
-  new FacebookStrategy(
-    {
-      clientID: config.facebook.id!,
-      clientSecret: config.facebook.secret!,
-      callbackURL: `${config.domainName}/login/facebook/callback`,
-      profileFields: ['id', 'email', 'first_name', 'last_name'],
-      passReqToCallback: true
-    },
-    (req, accessToken, refreshToken, profile, done) => {
-      if (!profile) {
-        return done(null, null);
-      }
-
-      const email = profile.emails![0].value;
-
-      User.findOne({
-        $or: [{ email: email }, { providers: { facebook: profile.id } }]
-      })
-        .then((user) => {
-          if (!user) {
-            user = new User();
-          }
-
-          if (!user.providers) {
-            user.providers = {};
-          }
-
-          user.email = email;
-          user.firstName = profile.name!.givenName;
-          user.lastName = profile.name!.familyName;
-          user.providers.facebook = profile.id;
-
-          user
-            .save()
-            .then((user) => {
-              new UserLogin({
-                userId: user._id,
-                accessToken: accessToken,
-                refreshToken: refreshToken,
-                createDate: new Date(),
-                remoteIp: getIp(req)
-              })
-                .save()
-                .then((login) => {
-                  user.login = login;
-                  return done(null, user);
-                })
-                .catch((err) => {
-                  console.error(err);
-                  return done(err, null);
-                });
-            })
-            .catch((err) => {
-              console.error(err);
-              return done(err, null);
-            });
-        })
-        .catch((err) => {
-          console.error(err);
-          return done(err, null);
-        });
-    }
-  )
-);
-
-passport.use(
-  new GoogleStrategy(
-    {
-      clientID: config.google.id!,
-      clientSecret: config.google.secret!,
-      callbackURL: `${config.domainName}/login/google/callback`,
-      passReqToCallback: true
-    },
-    (req, accessToken, refreshToken, profile, done) => {
-      if (profile == null) {
-        return done(null, null);
-      }
-
-      const email = profile.emails![0].value;
-
-      User.findOne({
-        $or: [{ email: email }, { providers: { google: profile.id } }]
-      })
-        .then((user) => {
-          if (user == null) {
-            user = new User();
-          }
-          if (user.providers == null) {
-            user.providers = {};
-          }
-
-          user.email = email;
-          user.firstName = profile.name!.givenName;
-          user.lastName = profile.name!.familyName;
-          user.providers.google = profile.id;
-
-          user
-            .save()
-            .then((user) => {
-              new UserLogin({
-                userId: user._id,
-                accessToken: accessToken,
-                refreshToken: refreshToken,
-                createDate: new Date(),
-                remoteIp: getIp(req)
-              })
-                .save()
-                .then((login) => {
-                  user.login = login;
-                  return done(null, user);
-                })
-                .catch((err) => {
-                  console.error(err);
-                  return done(err, null);
-                });
-            })
-            .catch((err) => {
-              console.error(err);
-              return done(err, null);
-            });
-        })
-        .catch((err) => {
-          console.error(err);
-          return done(err, null);
-        });
-    }
-  )
-);
-
 const middlewares = [];
 
 if (process.env.NODE_ENV === 'development') {
@@ -200,15 +69,6 @@ if (process.env.NODE_ENV === 'development') {
   cookieSession.cookie!.secure = true;
 }
 
-//define REST proxy options based on logged in user
-passport.serializeUser((user, done) => {
-  done(null, user);
-});
-
-passport.deserializeUser((obj, done) => {
-  done(null, obj);
-});
-
 middlewares.push(
   cookieParser(),
   session(cookieSession),
@@ -220,7 +80,7 @@ middlewares.push(
 );
 
 const app: express.Application = createExpressServer({
-  controllers: [RequestsController],
+  controllers: [LoginController, RequestsController],
   middlewares
 });
 const server = require('http').Server(app);
@@ -358,40 +218,6 @@ app.put('/api/makers/:id', (req: express.Request, res: express.Response) => {
     return res.send(result);
   });
 });
-
-app.get(
-  '/login/facebook',
-  passport.authenticate('facebook', { scope: 'email' })
-);
-
-app.get(
-  '/login/facebook/callback',
-  passport.authenticate('facebook', {
-    successRedirect: '/',
-    failureRedirect: '/login'
-  }),
-  (req: express.Request, res: express.Response) => {
-    console.log('facebook logged in');
-    res.send('Logged In.');
-  }
-);
-
-app.get(
-  '/login/google',
-  passport.authenticate('google', { scope: ['profile', 'email'] })
-);
-
-app.get(
-  '/login/google/callback',
-  passport.authenticate('google', {
-    successRedirect: '/',
-    failureRedirect: '/login'
-  }),
-  (req: express.Request, res: express.Response) => {
-    console.log('google logged in');
-    res.send('Logged In.');
-  }
-);
 
 // The "catchall" handler: for any request that doesn't
 // match one above, send back React's index.html file.
