+++ server/controllers/index.ts
@@ -1,1 +1,2 @@
+export { default as LoginController } from './login';
 export { default as RequestsController } from './requests';
